Work on "Shelf" abstraction and shortlist

+++ ocher/ux/fb/RenderFb.cpp
@@ -268,8 +268,8 @@ int RenderFb::render(Pagination* pagination, unsigned int pageNum, bool doBlit)
 #ifdef CPS_STATS
                         chars += str->length();  // miscounts UTF8...
 #endif
-                        clc::Log::debug(LOG_NAME, "output (%d pts%s%s) %d bytes", a[ai].pts, a[ai].b?" bold":"",
-                                a[ai].em?" italics":"", str->length());
+                        clc::Log::debug(LOG_NAME, "output (%d pts%s%s) %u bytes", a[ai].pts, a[ai].b?" bold":"",
+                                a[ai].em?" italics":"", (unsigned int)str->length());
                         int breakOffset = outputWrapped(str, strOffset, doBlit);
                         strOffset = 0;
                         if (breakOffset >= 0) {
